Add files via upload
This is only the source code and you will have to download the mixer music files and pictures mentioned in the code to run the game file properly. if you wish to change the BGM Musi c PICTURE or resolution then you may do so in your own source code and you will have to change the name of pictures and BGM  of the source code if you wish to change the ingame Music and Pictures.